Bug 898936 - Fix some build warnings on ARM gcc; r=mjrosenb

+++ js/src/jit/PerfSpewer.cpp
@@ -149,7 +149,7 @@ PerfSpewer::writeProfile(JSScript *script,
         if (size > 0) {
             fprintf(fp_,
                     "%lx %lx %s:%d: Func%02d\n",
-                    reinterpret_cast<uintptr_t>(code->raw()),
+                    reinterpret_cast<unsigned long>(code->raw()),
                     size,
                     script->filename(),
                     script->lineno,
@@ -170,7 +170,8 @@ PerfSpewer::writeProfile(JSScript *script,
             if (cur < blockStart) {
                 fprintf(fp_,
                         "%lx %lx %s:%d: Func%02d-Block?\n",
-                        cur, blockStart - cur,
+                        static_cast<unsigned long>(cur),
+                        static_cast<unsigned long>(blockStart - cur),
                         script->filename(), script->lineno,
                         thisFunctionIndex);
             }
@@ -181,7 +182,7 @@ PerfSpewer::writeProfile(JSScript *script,
             if (size > 0) {
                 fprintf(fp_,
                         "%lx %lx %s:%d:%d: Func%02d-Block%d\n",
-                        blockStart, size,
+                        static_cast<unsigned long>(blockStart), size,
                         r.filename, r.lineNumber, r.columnNumber,
                         thisFunctionIndex, r.id);
             }
@@ -193,7 +194,8 @@ PerfSpewer::writeProfile(JSScript *script,
         if (cur < funcEnd) {
             fprintf(fp_,
                     "%lx %lx %s:%d: Func%02d-OOL\n",
-                    cur, funcEnd - cur,
+                    static_cast<unsigned long>(cur),
+                    static_cast<unsigned long>(funcEnd - cur),
                     script->filename(), script->lineno,
                     thisFunctionIndex);
         }
